+++ src/main/java/br/com/aviapp/api/domain/factories/AviaryReportFactory.java
@@ -2,10 +2,9 @@ package br.com.aviapp.api.domain.factories;
 
 import br.com.aviapp.api.domain.entities.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
-public class AviaryReportFactory extends CollectCalculator{
+public class AviaryReportFactory extends CollectCalculator {
 
     public static AviaryReportBO createFromAviary(AviaryBO aviary, List<CollectEggBO> collectEggs, List<CollectChickenBO> collectChickens) {
         int totalEggsCollected = calculateTotalEggs(collectEggs);
@@ -16,11 +15,11 @@ public class AviaryReportFactory extends CollectCalculator{
         int currentChickens = aviary.getCurrentAmountOfChickens();
         int currentRoosters = aviary.getCurrentAmountOfRoosters();
         int totalBirds = currentChickens + currentRoosters;
-        BigDecimal production = BigDecimal.valueOf(totalEggsCollected / currentChickens);
-        BigDecimal roosterMortality= BigDecimal.valueOf(totalDeadRoosters / currentRoosters);
-        BigDecimal chickenMortality = BigDecimal.valueOf(totalDeadChickens / currentChickens);
-        BigDecimal mortality = BigDecimal.valueOf(totalDeadBirds / totalBirds);
-        BigDecimal chickenRoosterProportion = BigDecimal.valueOf(currentChickens / currentRoosters);
+        double production = (double) totalEggsCollected / currentChickens;
+        double roosterMortality = (double) totalDeadRoosters / currentRoosters;
+        double chickenMortality = (double) totalDeadChickens / currentChickens;
+        double mortality = (double) totalDeadBirds / totalBirds;
+        double chickenRoosterProportion = (double) currentChickens / currentRoosters;
 
         List<EggDetailBO> quantityByEggType = calculateEggsByType(collectEggs);
